Extracted mclose.
